Docs: Remove installation from docs but keep relevant sections
Removes documentation for installation since the supported installation
is via rpm and presto-admin.  Moves documentation for jdbc and cli to a
section on querying and documentation of the configuration to a
configuration section.  Also updates the configuration documentation
with the correct paths to the configuration directory for an rpm
install.  Fully removes documentation for verifier and
benchmark-driver

Task: SWARM-549