Allow LIKE predicates in join condition
So far only comparison predicates (=, !=, <, ...) were allowed as
conjuncts in condition clause for inner joins. This patch extends the
list of allowed expressions with LIKE expression.

There is limitation that all columns on values side of join must
correspond to one of joined tables and all columns on pattern side of
join must correspond to the other. This is consequence of currently
existent limitation and should be fixed holistically for both
ComparisonExpression and LikePredicate based joins in separate patch.