Support non correlated scalar subqueries.
Scalar subqueries are rewritten into a join in a following manner.
From:
 `SELECT * FROM nation WHERE regionkey = (
 SELECT regionkey FROM region WHERE name = 'AFRICA'
 )`
into:
 `SELECT * FROM nation, (
 SELECT regionkey FROM region WHERE name = 'AFRICA'
 ) temp
 WHERE nation.regionkey = temp.regionkey`

Also above subquery plan is wrapped in a ScalarNode which goal is to
maintain scalar subquery returns at most single value. To achieve that
during plan fragmentation, there is a plan fragment generated
from ScalarNode so all the results of the subquery goes through one
node.

Fixes #2878.